+++ wutengine/src/builtins/components/util/framerate.rs
@@ -6,13 +6,14 @@ use wutengine_macro::system;
 
 use crate::command::Command;
 use crate::log;
+use crate::time::Time;
 
 /// Framerate counter component. Requires [framerate_counter_system] to be present.
 ///
 /// Every second, logs the average FPS and frametime of the last second with level `info`
 #[derive(Debug)]
 pub struct FramerateCounter {
-    prev_time: Instant,
+    prev_report_time: Instant,
     num_frames: Saturating<usize>,
 }
 
@@ -20,7 +21,7 @@ impl FramerateCounter {
     /// Creates a new [FramerateCounter]
     pub fn new() -> Self {
         Self {
-            prev_time: Instant::now(),
+            prev_report_time: Instant::now(),
             num_frames: Saturating(0),
         }
     }
@@ -43,9 +44,11 @@ pub fn framerate_counter_system(
 ) {
     counter.num_frames += 1;
 
-    let cur_time = Instant::now();
+    let cur_time = Time::get().frame_start;
 
-    let duration_since_last_report = cur_time.duration_since(counter.prev_time).as_secs_f64();
+    let duration_since_last_report = cur_time
+        .duration_since(counter.prev_report_time)
+        .as_secs_f64();
 
     if duration_since_last_report >= 1.0 {
         let fps = counter.num_frames.0 as f64 / duration_since_last_report;
@@ -53,6 +56,6 @@ pub fn framerate_counter_system(
         log::info!("FPS: {}, avg. frametime: {}ms", fps, (1.0 / fps) * 1000.0);
 
         counter.num_frames = Saturating(0);
-        counter.prev_time = cur_time;
+        counter.prev_report_time = cur_time;
     }
 }